dashboard/requirements.txt
streamlit==1.31.1
pandas==1.5.3
plotly==5.18.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv